+++ src/DbTarget.php
@@ -18,13 +18,15 @@ final class DbTarget extends Target
 {
     /**
      * @param ConnectionInterface $db The database connection instance.
-     * @param string $table The name of the database table to store the log messages. Defaults to "log".
+     * @param string $table The name of the database table to store the log messages. Defaults to "{{%yii_log}}".
+     * @param string[] $levels The {@see \Psr\Log\LogLevel log message levels} that this target is interested in.
      */
     public function __construct(
         private readonly ConnectionInterface $db,
-        private readonly string $table = '{{%yii_log}}'
+        private readonly string $table = '{{%yii_log}}',
+        array $levels = []
     ) {
-        parent::__construct();
+        parent::__construct($levels);
     }
 
     /**