+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.1 under development
 
+- New #66: Add optional `$levels` parameter to `DbTarget` constructor allowing log level filtering at instantiation (@samdark)
 - Chg #74: Bump minimal required PHP version to 8.1 (@batyrmastyr)
 
 ## 1.0.0 May 09, 2023